go/consensus/tendermint: Move transaction logic to its own file
